load_balancer to node_balancer
The keyword for creating a load balancer has been updated from `load_balancer` to `node_balancer`. Apparently this is Linode's term for load balancers.
I realise that this is technically an breaking change, but this has come about due to my tests on the API and the load balancer feature not working. As far as i can tell it never did work as previously documented. I can confirm that the `node_balancer` term does function as intended.